SOLID-ize printer pt. 3 (completely rid BoardPrinter class)
