+++ src/main/java/com/jesseoj98/tictactoe/util/printer/board/actual/ActualCoordinateBoardPrinter.java
@@ -1,4 +1,4 @@
-package com.jesseoj98.tictactoe.util.printer.board;
+package com.jesseoj98.tictactoe.util.printer.board.actual;
 
 import com.jesseoj98.tictactoe.domain.GameBoard;
 import com.jesseoj98.tictactoe.util.helper.occupier.BoardOccupier;
@@ -6,7 +6,7 @@ import com.jesseoj98.tictactoe.util.helper.validator.printer.PrinterValidator;
 import com.jesseoj98.tictactoe.util.printer.board.structure.BoardStructurePrinter;
 import com.jesseoj98.tictactoe.util.printer.space.SpacePrinter;
 
-public class BoardPrinter {
+public class ActualCoordinateBoardPrinter {
 
     private static final BoardStructurePrinter bsp = new BoardStructurePrinter();
 
@@ -16,33 +16,13 @@ public class BoardPrinter {
 
     private static final SpacePrinter sp = new SpacePrinter();
 
-    public void printSimpleCoordinateGameBoard(char[] gameBoard) {
-        bsp.printBoardDivider(0, 3);
-        printSimpleBoard(gameBoard);
-    }
-
-    private void printSimpleBoard(char[] gameBoard) {
-        for (int i = 0; i < GameBoard.GAME_BOARD_SPACES; i++) {
-            if (pv.endOfSimpleRowMatch(i)) {
-                System.out.println(" " + sp.printSpace(gameBoard[i], i) + " |");
-                bsp.printBoardDivider(0, 3);
-            } else {
-                if (i % 3 != 0) {
-                    System.out.print(sp.printSpace(gameBoard[i], i) + " |");
-                } else {
-                    System.out.print("| " + sp.printSpace(gameBoard[i], i) + " | ");
-                }
-            }
-        }
-    }
-
-    public void printCoordinates(char[][] gameBoard) {
+    public void printActualBoardSequence(char[][] gameBoard) {
         bsp.printNumbers(4, 3);
-        printCoordinatesGameBoard(gameBoard);
+        printActualBoard(gameBoard);
         bsp.printNumbers(4, 3);
     }
 
-    private void printCoordinatesGameBoard(char[][] gameBoard) {
+    private void printActualBoard(char[][] gameBoard) {
         bsp.printBoardDivider(0, 3);
         for (int i = 0; i < GameBoard.GAME_BOARD_DIMENSION; i++) {
             System.out.print((i + 1) + " ");
@@ -60,5 +40,4 @@ public class BoardPrinter {
         }
         bsp.printBoardDivider(0, 3);
     }
-
 }
